backend/lambda/agent-single/requirements.txt
boto3==1.34.39
pydantic==2.7.1
requests==2.31.0
openai==1.14.2
onnxruntime==1.16.3
Pillow==10.2.0

